Stub out the GRPC Terraform provider server.
Create a Server type and stub out the methods necessary to fill the
ProviderServer interface.

This involved copying over the `tfplugin5` package from
terraform-plugin-sdk. That's not ideal; we're going to have to figure
out a solution using type aliases or a separate Go Module at some point.
But for now, it lets us prototype.