Finish prototyping.
Implement actual protocol functions and add a resource muxer type to
route resource-level requests based on which resource the request is
for.

Add a helper NewResourceServer function that will take a server to use
by default, a server to use for specific resources, and the resources to
use the alternate server for.

Add an example of using NewResourceServer, which is how most providers
will likely interact with this package. We can't actually show the full
usage, because we have a type system problem.

The tfplugin5 internal package we're using in our type signatures is
github.com/hashicorp/terraform-plugin-mux/internal/tfplugin5, but the
tfplugin5 package the SDK is using is
github.com/hashicorp/terraform-plugin-sdk/internal/tfplugin5, which
means all the types are not considered the same. So our Server
implementation doesn't actually implement the
github.com/hashicorp/terraform-plugin-sdk/plugin#ProviderServer
interface, because the types are different. We need to resolve this by
pointing both at the same package, or making one's types type aliases of
the other's.