Add tests, update behavior.
Enforce equality for schemas returned from servers for provider and
provider_meta, in accordance with our docs and RFC. This is better than
requiring one provider to set it, because each provider will be parsing
these values, so they should all agree on what they are.

Add a bunch of tests.