Refactor logging into shared internal package
Reference: https://github.com/hashicorp/terraform-plugin-mux/issues/19
Reference: https://github.com/hashicorp/terraform-plugin-mux/issues/33

In the near future, this project will be implementing multiple packages which will be reliant on shared logging details. Refactoring this code now will prevent duplication when those packages get created.

This logging package structure is similar to what is being implemented in terraform-plugin-go, not only for code consistency, but in the event that portions or all of this project is subsumed into that project.