tf5muxserver+tf6muxserver: Remove resource schema caching
Reference: https://github.com/hashicorp/terraform-plugin-mux/issues/168

This change removes the `GetProviderSchema` RPC-based caching logic in `NewMuxServer` and directly calls the underlying providers within that RPC. To fully support the removal of resource schema caching, the `PlanResourceChange` logic for checking the `ProposedNewState` data for a destroy plan was optimized to no longer require decoding the data with the type information.

For larger scale providers (either number of resources and/or resources with large schemas), this should reduce resident memory utilization with little performance penalty.